Moves integration tests into ignite-cli
