+++ packages/ignite-cli/tests/fast/cli/cliConfig.test.js
@@ -17,7 +17,7 @@ mockery.registerMock('gluegun', {
 })
 
 // our cli
-const cli = require('../../src/cli/cli')
+const cli = require('../../../src/cli/cli')
 
 test('ignite', async t => {
   mockGluegunBuilder.onCreateRuntime(builderProps => {
