+++ packages/ignite-cli/tests/fast/cli/cliError.test.js
@@ -20,7 +20,7 @@ mockery.registerMock('gluegun', {
 })
 
 // our cli
-const cli = require('../../src/cli/cli')
+const cli = require('../../../src/cli/cli')
 
 test('ignite', async t => {
   const mockGluegunRunContext = {
