+++ packages/ignite-cli/tests/fast/fixtures/ignite-valid-plugin/package.json
