+++ packages/ignite-cli/tests/integration/ignite-cli/badCommands.test.js
