+++ packages/ignite-cli/tests/integration/ignite-cli/noArgs.test.js
