+++ packages/ignite-cli/tests/integration/ignite-cli/version.test.js
