+++ packages/ignite-cli/tests/integration/ignite-new/newInvalid.test.js
