--- packages/ignite-integration-tests/.gitignore
@@ -1,1 +1,0 @@
-out
