--- packages/ignite-integration-tests/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "ignite-integration-tests",
-  "private": true,
-  "version": "0.0.1",
-  "description": "A place to run some long running integration tests.",
-  "scripts": {
-    "lint": "standard",
-    "integration": "ava -s",
-    "watch": "ava -s --watch"
-  },
-  "license": "MIT",
-  "devDependencies": {
-    "ava": "^0.18.1",
-    "babel-eslint": "^7.1.1",
-    "standard": "^8.6.0"
-  },
-  "dependencies": {
-    "execa": "^0.6.0",
-    "fs-jetpack": "^0.11.0",
-    "ramda": "^0.23.0",
-    "ramdasauce": "^1.2.0",
-    "shelljs": "^0.7.6"
-  },
-  "files": [
-    "readme.md",
-    "test"
-  ],
-  "repository": "infinitered/ignite",
-  "author": {
-    "name": "Infinite Red",
-    "email": "npm@infinite.red",
-    "url": "https://github.com/infinitered/ignite"
-  },
-  "standard": {
-    "parser": "babel-eslint",
-    "globals": [
-      "__DEV__"
-    ]
-  }
-}
