--- packages/ignite-integration-tests/readme.md
@@ -1,8 +1,0 @@
-# ignite-integration-tests
-
-These are long running end-to-end integration tests.
-
-## Usage
-
-$ yarn run integration
-
